# docs/handoff/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SSS Artifact Relay API
  version: 1.0.0
  summary: Ephemeral authenticated store-and-forward artifact relay
servers:
  - url: https://drop.example.com
security:
  - basicAuth: []
paths:
...
  /local/r/{code}:
    get:
      security: []
      summary: Return an existing VPS-local payload path
      description: Unix-socket listener only. Must not be exposed by the public listener.
      parameters:
        - $ref: "#/components/parameters/CodePath"
      responses:
        "200":
          description: Plain path or JSON local claim
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                $ref: "#/components/schemas/LocalClaim"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Expired"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    claimBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    CodePath:
      name: code
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Code"
...
  responses:
...
    NotFound:
      description: Transfer or resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
        text/plain:
          schema:
            type: string
    Expired:
      description: Transfer or claim expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
        text/plain:
          schema:
            type: string
...
  schemas:
    Code:
      type: string
      pattern: "^[0-9A-HJKMNP-TV-Z]{4}-?[0-9A-HJKMNP-TV-Z]{4}$"
      examples: ["K7M4-Q2PX"]
...
    LocalClaim:
      type: object
      required: [ok, code, path, read_only, lease_until]
      properties:
        ok:
          type: boolean
          const: true
        code:
          $ref: "#/components/schemas/Code"
        path:
          type: string
        read_only:
          type: boolean
          const: true
        lease_until:
          type: string
          format: date-time
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"